+++ tests/Feature/Translation/TranslationIndexTest.php
@@ -24,7 +24,7 @@ final class TranslationIndexTest extends TestCase
         );
         $declaration = new TranslationDeclaration('article.body', 'messages', 'en', 'Body', 'file:///translations/messages.en.yaml', $this->range());
         $reference = new TranslationReference('article.title', 'messages', 'file:///src/Controller.php', $this->range());
-        $index->replaceSources(new TranslationSourceFacts('file:///source', [$declaration], [$reference]));
+        $index->replaceSources(new TranslationSourceFacts('file:///source', [$declaration], [$reference], ['%current_domain_name%']));
 
         self::assertCount(2, $index->messages('messages', 'article.title'));
         self::assertSame([$declaration], $index->declarations('messages', 'article.body'));
@@ -32,6 +32,7 @@ final class TranslationIndexTest extends TestCase
         self::assertSame(['article.body', 'article.title'], $index->keys('messages', 'article.'));
         self::assertSame(['admin', 'messages'], $index->domains());
         self::assertSame(['en', 'fr'], $index->locales());
+        self::assertSame(['current_domain_name'], $index->globalParameters());
         self::assertTrue($index->isComplete());
     }
 
@@ -43,9 +44,10 @@ final class TranslationIndexTest extends TestCase
         $index->replaceSources(new TranslationSourceFacts('file:///source', [$first]));
         self::assertSame(['first'], $index->keys('messages', ''));
 
-        $index->overlay(new TranslationSourceFacts('file:///source', [$second]));
+        $index->overlay(new TranslationSourceFacts('file:///source', [$second], [], [], true));
         self::assertSame(['second'], $index->keys('messages', ''));
         self::assertSame([], $index->declarations('messages', 'first'));
+        self::assertNull($index->globalParameters());
 
         $index->replaceRuntime(false, new TranslationMessage('runtime', 'runtime', 'de', 'Runtime'));
         self::assertSame(['messages', 'runtime'], $index->domains());